Moved everything around and ZmqClient is now ugly
Want to create an "empty" ZmqClient on launch that is only "really"
initialized when someone actually wants to do that. Like when the user
clicks on a camera launcher. Otherwise, why are we loading the camera?

Also, I think I was creating a new ZMQ connection every frame??? I can
see why it wasn't working... :o